ajout de la page d'accueil sous forme de carte
