+++ app/src/main/res/drawable/round_shape.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content">
+    <corners android:radius="12dp"/>
+    <solid android:color="#00ffffff"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content" />
+    <padding android:left="8dp"
+        android:top="8dp"
+        android:right="8dp"
+        android:bottom="8dp"/>
+    <corners android:radius="12dp"/>
+    <stroke android:color="#FFFFFFFF"
+        android:width="12dp"/>
+   </shape>
