+++ app/src/main/res/layout/row_products_main.xml
@@ -1,0 +1,70 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+    <android.support.v7.widget.CardView
+        xmlns:card_view="http://schemas.android.com/apk/res-auto"
+        android:layout_height="180dp"
+        android:layout_width="370dp"
+        android:layout_alignParentTop="true"
+        android:layout_centerHorizontal="true"
+        card_view:cardCornerRadius="5dp"
+        android:layout_margin="6dp"
+        android:onClick="openRecipe"
+        android:id="@+id/view">
+        <TextView
+            android:layout_height="30dp"
+            android:text="Vos produits"
+            android:background="#77FFF9C4"
+            android:layout_width="match_parent"
+            android:layout_margin="2dp"
+            android:gravity="center"/>
+        <FrameLayout android:layout_height="80dp"
+            android:layout_marginTop="50dp"
+            android:layout_width="80dp"
+            android:layout_gravity="left">
+            <ImageView
+                android:src="@drawable/banana"
+                android:layout_width="80dp"
+                android:layout_height="80dp"
+                android:layout_gravity="center"
+                />
+            <ImageView
+                android:layout_width="85dp"
+                android:layout_height="85dp"
+                android:src="@drawable/round_shape"
+                android:layout_gravity="center"/>
+        </FrameLayout>
+        <FrameLayout android:layout_height="80dp"
+            android:layout_width="80dp"
+            android:layout_gravity="center">
+            <ImageView
+                android:src="@drawable/steak"
+                android:layout_width="80dp"
+                android:layout_height="80dp"
+                android:layout_gravity="center"
+                />
+            <ImageView
+                android:layout_width="85dp"
+                android:layout_height="85dp"
+                android:src="@drawable/round_shape"
+                android:layout_gravity="center"/>
+        </FrameLayout>
+        <FrameLayout android:layout_height="80dp"
+            android:layout_width="80dp"
+            android:layout_gravity="right"
+            android:layout_marginTop="50dp">
+            <ImageView
+                android:src="@drawable/yogourt"
+                android:layout_width="78dp"
+                android:layout_height="78dp"
+                android:layout_gravity="center"
+                />
+            <ImageView
+                android:layout_width="85dp"
+                android:layout_height="85dp"
+                android:src="@drawable/round_shape"
+                android:layout_gravity="center"/>
+        </FrameLayout>
+    </android.support.v7.widget.CardView>
+</RelativeLayout>
