+++ app/src/main/res/layout/row_recipe_main.xml
@@ -1,0 +1,43 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+    <android.support.v7.widget.CardView
+        xmlns:card_view="http://schemas.android.com/apk/res-auto"
+        android:layout_height="180dp"
+        android:layout_width="370dp"
+        android:layout_alignParentTop="true"
+        android:layout_centerHorizontal="true"
+        card_view:cardCornerRadius="5dp"
+        android:layout_margin="6dp"
+        >
+        <FrameLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent">
+            <ImageView
+                android:layout_width="392dp"
+                android:layout_height="220dp"
+                android:scaleType="fitCenter"
+                android:adjustViewBounds="true"
+                android:src="@drawable/recipe"/>
+
+            <TextView
+                android:layout_width="230dp"
+                android:layout_height="50dp"
+                android:text="@string/todays_recipe"
+                android:gravity="center"
+                android:background="#CCDDDDDD"
+                android:layout_gravity="left|top" />
+
+            <TextView
+                android:layout_width="230dp"
+                android:layout_height="70dp"
+                android:text="Filet de poisson et omelette panée a la frangipane de france"
+                android:padding="6dp"
+                android:background="#DDBBBBBB"
+                android:layout_gravity="right|bottom"
+                android:gravity="center"/>
+        </FrameLayout>
+    </android.support.v7.widget.CardView>
+
+</RelativeLayout>
